Move NormalizeURL into its own function
